extrat plot conversion from chatSideEffect to a separate file
